+++ libraries/Prism/CHANGELOG.md
@@ -1,6 +1,13 @@
 Prism Library Changelog
 ==========================
 
+###v1.17.4
+* Added [LoadingOverlay](http://gasparesganga.com/labs/jquery-loading-overlay/) to UI library.
+
+###v1.17.3
+* Fixed Prism\File\Image::crop().
+* Improved Prism\Utilities\NetworkHelper::isValidIp().
+
 ###v1.17.2
 * It was made compatible with PHP 5.5.